# src/main/resources/partner-gateway-openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Globex Partners API Gateway",
        "version": "1.0.0",
        "description": "Globex APIs made accessible to global partners to view Globex' catalog and products"
    },
    "paths": {
        "/partners/services/catalog/product": {
            "summary": "Get a paginated list of products",
            "description": "Get a paginated list of products",
            "get": {
                "responses": {
                    "200": {
                        "description": "Product list found."
                    }
                },
                "description": "Get a paginated list of products"
            },
            "parameters": [
                {
                    "name": "page",
                    "description": "The page number. First page is 1.",
                    "schema": {
                        "type": "integer"
                    },
                    "in": "query"
                },
                {
                    "name": "limit",
                    "description": "The maximum number of products returned.",
                    "schema": {
                        "type": "integer"
                    },
                    "in": "query"
                },
                {
                    "name": "inventory",
                    "description": "Whether to include inventory information.",
                    "schema": {
                        "type": "boolean"
                    },
                    "in": "query"
                }
            ]
        },
        "/partners/services/catalog/product/{id}": {
            "summary": "Get product by id",
            "description": "Get product by id",
            "get": {
                "parameters": [
                    {
                        "name": "inventory",
                        "description": "If true, include inventory information",
                        "schema": {
                            "type": "boolean"
                        },
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "summary": "Get product by id",
                "description": "Get product by id"
            },
            "parameters": [
                {
                    "name": "id",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/partners/services/catalog/product/list/{ids}": {
            "summary": "Get product list by ids",
            "description": "Get product list by ids",
            "get": {
                "parameters": [
                    {
                        "name": "inventory",
                        "description": "Whether to include inventory information",
                        "schema": {
                            "type": "boolean"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product list found"
                    }
                },
                "summary": "Get product list",
                "description": "Get product list"
            },
            "parameters": [
                {
                    "name": "ids",
                    "description": "comma separated list of product ids",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/partners/services/catalog/product/category/{categories}": {
            "summary": "Get products by categories",
            "description": "Return the list of products that belong to the given categories.",
            "get": {
                "responses": {
                    "200": {
                        "description": "Product list found."
                    }
                },
                "summary": "Get products by category"
            },
            "parameters": [
                {
                    "name": "categories",
                    "description": "Comma separated list of category names.",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                },
                {
                    "name": "inventory",
                    "description": "Whether to include inventory information.",
                    "schema": {
                        "type": "boolean"
                    },
                    "in": "query"
                }
            ]
        },
        "/partners/services/catalog/product/tag/{tags}": {
            "summary": "Get products by tags",
            "description": "Return the list of products that are tagged with he given tags.",
            "get": {
                "responses": {
                    "200": {
                        "description": "Product list found."
                    }
                },
                "summary": "Get products by tag"
            },
            "parameters": [
                {
                    "name": "inventory",
                    "description": "Whether to include inventory information.",
                    "schema": {
                        "type": "boolean"
                    },
                    "in": "query"
                },
                {
                    "name": "tags",
                    "description": "Comma separated list of tags",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        }
    }
}